requirements.txt
streamlit
openai>=1.0.0
sentence-transformers
faiss-cpu
